[elf] E: Deduplicate ELF32 types
Extract duplicated ELF32 constants, structures, and validation logic
from kernel, uservm, and nvx into a shared elf::elf32 module. Gate the
goblin-dependent relocation code behind a "relocation" feature flag so
consumers that only need ELF32 definitions have zero extra dependencies.

- Add elf::elf32 module with shared constants, Elf32Fhdr, Elf32Phdr,
  Elf32Dyn, and Elf32Rel types.
- Add Elf32Fhdr::validate() for common header checks (magic, class,
  encoding, version, phentsize) and Elf32Phdr::is_loadable()/validate()
  for segment checks.
- Move relocation types (SymbolTable, StringTable, RelocationTable,
  RelocationType, etc.) into a feature-gated relocation module.
- Make goblin, num_enum, and static_assert optional dependencies of elf,
  required only with the "relocation" feature.
- Update kernel, uservm, nvx, and syscall crates to use the shared
  module, removing ~400 lines of triplicated definitions.
