+++ src/libs/nvx/Cargo.toml
@@ -14,6 +14,7 @@ homepage.workspace = true
 alloc = { version = "1.0.1", optional = true, package = "rustc-std-workspace-alloc" }
 cfg-if = { workspace = true }
 config = { workspace = true }
+elf = { workspace = true }
 syslog = { workspace = true, default-features = true }
 sys = { workspace = true, features = ["kcall"] }
 core = { version = "1.0.1", optional = true, package = "rustc-std-workspace-core" }
