+++ src/uservm/Cargo.toml
@@ -24,6 +24,7 @@ serde_cbor = { workspace = true }
 cfg-if = { workspace = true }
 config = { workspace = true }
 control-plane-api = { workspace = true }
+elf = { workspace = true }
 hyperlight-host = { workspace = true, optional = true }
 log = { workspace = true }
 multibin = { workspace = true, features = ["std"] }
